Fix issue #547 'Deleting bucket after inserting a nil value'
Looks like any key with a nil value would be considered as a sub
bucket, replaced the previous 'v == nil' test by a more explicit test
on flags. In the general case it should do the same (a sub-bucket does
not have a non-nil value, does it?) but it should also handle properly
the case of nil values on real keys.